Put strat matching in subdirectory
