+++ .idea/sqldialects.xml
@@ -4,9 +4,8 @@
     <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/ad-hoc-migrations/rotation-cache-cleanup.sql" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/database/migrations/update_macrostrat/03-primary-key-seq.sql" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/subsystems/paleogeography/corelle/fixtures/01-corelle-macrostrat.sql" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/subsystems/sgp/column-strat-names.sql" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/subsystems/sgp/get-strat-name-footprints.sql" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/subsystems/sgp/measurements-to-match.sql" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/subsystems/sgp/match/sql/column-strat-names.sql" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/cli/macrostrat/cli/subsystems/sgp/match/sql/measurements-to-match.sql" dialect="PostgreSQL" />
     <file url="PROJECT" dialect="PostgreSQL" />
   </component>
 </project>
