+++ cli/macrostrat/cli/subsystems/sgp/match/sql/column-strat-names.sql
