--- cli/macrostrat/cli/subsystems/sgp/small-data-fixes.sql
@@ -1,3 +1,0 @@
--- Small data fixes that should be checked in as a migration script.
--- or held in a separate file for eventual application to the production database.
-
